FitLife-Nutrition-AI/requirements.txt
<<<<<<< HEAD


streamlit
plotly
scikit-learn
pandas
numpy

# RAG Dependencies
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
langchain==0.0.340
langchain-community==0.0.10
faiss-cpu==1.7.4
ollama>=0.5.4,<0.7
pyyaml==6.0.1
=======
streamlit>=1.28.0
plotly>=5.17.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
>>>>>>> 6660fe86de3a80c1cfe400a7d0c97cbe914de13a
